created context to discount, sales and voucher - create, edit, list, delete

+++ saleor/my_customs/merchant_dashboard/discount/views.py
@@ -9,6 +9,7 @@ from django.shortcuts import get_object_or_404, redirect
 from django.template.response import TemplateResponse
 from django.utils.translation import pgettext_lazy
 
+from saleor.my_customs.merchant.models import Merchant
 from . import forms
 from ....core.utils import get_paginator_items
 from ....discount import VoucherType
@@ -35,7 +36,14 @@ def get_voucher_type_forms(voucher, data):
 @staff_member_required
 @permission_required('discount.manage_discounts')
 def sale_list(request):
-    sales = Sale.objects.prefetch_related('products').order_by('name')
+    user = request.user
+    if not user.is_authenticated:
+        return  # todo: return unauthorized page
+    merchant = Merchant.objects.get_merchant_of_user(user)
+    if not merchant:
+        return TemplateResponse(request, 'my_customs/merchant_dashboard/not_registered.html')
+
+    sales = Sale.objects.get_by_merchant(merchant).prefetch_related('products').order_by('name')
     sale_filter = SaleFilter(request.GET, queryset=sales)
     sales = get_paginator_items(
         sale_filter.qs, settings.DASHBOARD_PAGINATE_BY,
@@ -49,13 +57,20 @@ def sale_list(request):
 @staff_member_required
 @permission_required('discount.manage_discounts')
 def sale_add(request):
+    user = request.user
+    if not user.is_authenticated:
+        return  # todo: return unauthorized page
+    merchant = Merchant.objects.get_merchant_of_user(user)
+    if not merchant:
+        return TemplateResponse(request, 'my_customs/merchant_dashboard/not_registered.html')
+
     sale = Sale()
-    form = forms.SaleForm(request.POST or None, instance=sale)
+    form = forms.SaleForm(request.POST or None, instance=sale, merchant = merchant)
     if form.is_valid():
         sale = form.save()
         msg = pgettext_lazy('Sale (discount) message', 'Added sale')
         messages.success(request, msg)
-        return redirect('dashboard:sale-update', pk=sale.pk)
+        return redirect('merchant_dashboard:sale-update', pk=sale.pk)
     ctx = {'sale': sale, 'form': form}
     return TemplateResponse(request, 'my_customs/merchant_dashboard/discount/sale/form.html', ctx)
 
@@ -63,13 +78,20 @@ def sale_add(request):
 @staff_member_required
 @permission_required('discount.manage_discounts')
 def sale_edit(request, pk):
-    sale = get_object_or_404(Sale, pk=pk)
-    form = forms.SaleForm(request.POST or None, instance=sale)
+    user = request.user
+    if not user.is_authenticated:
+        return  # todo: return unauthorized page
+    merchant = Merchant.objects.get_merchant_of_user(user)
+    if not merchant:
+        return TemplateResponse(request, 'my_customs/merchant_dashboard/not_registered.html')
+
+    sale = get_object_or_404(Sale.objects.get_by_merchant(merchant), pk=pk)
+    form = forms.SaleForm(request.POST or None, instance=sale, merchant=merchant)
     if form.is_valid():
         sale = form.save()
         msg = pgettext_lazy('Sale (discount) message', 'Updated sale')
         messages.success(request, msg)
-        return redirect('dashboard:sale-update', pk=sale.pk)
+        return redirect('merchant_dashboard:sale-update', pk=sale.pk)
     ctx = {'sale': sale, 'form': form}
     return TemplateResponse(request, 'my_customs/merchant_dashboard/discount/sale/form.html', ctx)
 
@@ -77,13 +99,20 @@ def sale_edit(request, pk):
 @staff_member_required
 @permission_required('discount.manage_discounts')
 def sale_delete(request, pk):
-    instance = get_object_or_404(Sale, pk=pk)
+    user = request.user
+    if not user.is_authenticated:
+        return  # todo: return unauthorized page
+    merchant = Merchant.objects.get_merchant_of_user(user)
+    if not merchant:
+        return TemplateResponse(request, 'my_customs/merchant_dashboard/not_registered.html')
+
+    instance = get_object_or_404(Sale.objects.get_by_merchant(merchant), pk=pk)
     if request.method == 'POST':
         instance.delete()
         msg = pgettext_lazy(
             'Sale (discount) message', 'Removed sale %s') % (instance.name,)
         messages.success(request, msg)
-        return redirect('dashboard:sale-list')
+        return redirect('merchant_dashboard:sale-list')
     ctx = {'sale': instance}
     return TemplateResponse(
         request, 'my_customs/merchant_dashboard/discount/sale/modal/confirm_delete.html', ctx)
@@ -92,7 +121,14 @@ def sale_delete(request, pk):
 @staff_member_required
 @permission_required('discount.manage_discounts')
 def voucher_list(request):
-    vouchers = (Voucher.objects.prefetch_related('products', 'collections')
+    user = request.user
+    if not user.is_authenticated:
+        return  # todo: return unauthorized page
+    merchant = Merchant.objects.get_merchant_of_user(user)
+    if not merchant:
+        return TemplateResponse(request, 'my_customs/merchant_dashboard/not_registered.html')
+
+    vouchers = (Voucher.objects.get_by_merchant(merchant).prefetch_related('products', 'collections')
                 .order_by('name'))
     voucher_filter = VoucherFilter(request.GET, queryset=vouchers)
     vouchers = get_paginator_items(
@@ -108,9 +144,16 @@ def voucher_list(request):
 @staff_member_required
 @permission_required('discount.manage_discounts')
 def voucher_add(request):
+    user = request.user
+    if not user.is_authenticated:
+        return  # todo: return unauthorized page
+    merchant = Merchant.objects.get_merchant_of_user(user)
+    if not merchant:
+        return TemplateResponse(request, 'my_customs/merchant_dashboard/not_registered.html')
+
     voucher = Voucher()
     type_base_forms = get_voucher_type_forms(voucher, request.POST)
-    voucher_form = forms.VoucherForm(request.POST or None, instance=voucher)
+    voucher_form = forms.VoucherForm(request.POST or None, instance=voucher, merchant = merchant)
     if voucher_form.is_valid():
         voucher_type = voucher_form.cleaned_data.get('type')
         form_type = type_base_forms.get(voucher_type)
@@ -119,11 +162,13 @@ def voucher_add(request):
             voucher = voucher_form.save()
         elif form_type.is_valid():
             voucher = form_type.save()
+            voucher.merchant = merchant
+            Voucher.save(voucher)
 
         if form_type is None or form_type.is_valid():
             msg = pgettext_lazy('Voucher message', 'Added voucher')
             messages.success(request, msg)
-            return redirect('dashboard:voucher-list')
+            return redirect('merchant_dashboard:voucher-list')
     ctx = {
         'voucher': voucher, 'default_currency': settings.DEFAULT_CURRENCY,
         'form': voucher_form, 'type_base_forms': type_base_forms}
@@ -134,9 +179,16 @@ def voucher_add(request):
 @staff_member_required
 @permission_required('discount.manage_discounts')
 def voucher_edit(request, pk):
-    voucher = get_object_or_404(Voucher, pk=pk)
+    user = request.user
+    if not user.is_authenticated:
+        return  # todo: return unauthorized page
+    merchant = Merchant.objects.get_merchant_of_user(user)
+    if not merchant:
+        return TemplateResponse(request, 'my_customs/merchant_dashboard/not_registered.html')
+
+    voucher = get_object_or_404(Voucher.objects.get_by_merchant(merchant), pk=pk)
     type_base_forms = get_voucher_type_forms(voucher, request.POST)
-    voucher_form = forms.VoucherForm(request.POST or None, instance=voucher)
+    voucher_form = forms.VoucherForm(request.POST or None, instance=voucher, merchant = merchant)
     if voucher_form.is_valid():
         voucher_type = voucher_form.cleaned_data.get('type')
         form_type = type_base_forms.get(voucher_type)
@@ -145,11 +197,13 @@ def voucher_edit(request, pk):
             voucher = voucher_form.save()
         elif form_type.is_valid():
             voucher = form_type.save()
+            voucher.merchant = merchant
+            Voucher.save(voucher)
 
         if form_type is None or form_type.is_valid():
             msg = pgettext_lazy('Voucher message', 'Updated voucher')
             messages.success(request, msg)
-            return redirect('dashboard:voucher-list')
+            return redirect('merchant_dashboard:voucher-list')
     ctx = {
         'voucher': voucher, 'default_currency': settings.DEFAULT_CURRENCY,
         'form': voucher_form, 'type_base_forms': type_base_forms}
@@ -160,13 +214,20 @@ def voucher_edit(request, pk):
 @staff_member_required
 @permission_required('discount.manage_discounts')
 def voucher_delete(request, pk):
-    instance = get_object_or_404(Voucher, pk=pk)
+    user = request.user
+    if not user.is_authenticated:
+        return  # todo: return unauthorized page
+    merchant = Merchant.objects.get_merchant_of_user(user)
+    if not merchant:
+        return TemplateResponse(request, 'my_customs/merchant_dashboard/not_registered.html')
+
+    instance = get_object_or_404(Voucher.objects.get_by_merchant(merchant), pk=pk)
     if request.method == 'POST':
         instance.delete()
         msg = pgettext_lazy(
             'Voucher message', 'Removed voucher %s') % (instance,)
         messages.success(request, msg)
-        return redirect('dashboard:voucher-list')
+        return redirect('merchant_dashboard:voucher-list')
     ctx = {'voucher': instance}
     return TemplateResponse(
         request, 'my_customs/merchant_dashboard/discount/voucher/modal/confirm_delete.html', ctx)
@@ -175,7 +236,14 @@ def voucher_delete(request, pk):
 @staff_member_required
 @permission_required('discount.manage_discounts')
 def ajax_voucher_list(request):
-    queryset = Voucher.objects.active(date=date.today())
+    user = request.user
+    if not user.is_authenticated:
+        return  # todo: return unauthorized page
+    merchant = Merchant.objects.get_merchant_of_user(user)
+    if not merchant:
+        return TemplateResponse(request, 'my_customs/merchant_dashboard/not_registered.html')
+
+    queryset = Voucher.objects.get_by_merchant(merchant).active(date=date.today())
 
     search_query = request.GET.get('q', '')
     if search_query:
